entity 수정중

+++ backend/src/main/kotlin/com/min/mockstock/infra/properties/UserAccountProperties.kt
@@ -1,8 +1,9 @@
 package com.min.mockstock.infra.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
+import java.math.BigDecimal
 
 @ConfigurationProperties(prefix = "user-account")
 data class UserAccountProperties(
-    val defaultBalance: Double
+    val defaultBalance: BigDecimal
 )
